+++ src/components/pages/Manage/Manage.tsx
@@ -9,6 +9,7 @@ import {
 } from '@src/state';
 import eventEmitter from '@src/utils/events';
 import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 import { doAntsRequireUpdate } from '../../../utils';
 import DomainsTable from '../../data-display/tables/DomainsTable';
@@ -16,6 +17,7 @@ import { RefreshIcon, SearchIcon } from '../../icons';
 import './styles.css';
 
 function Manage() {
+  const navigate = useNavigate();
   const [{ arioProcessId, aoNetwork, hyperbeamUrl }] = useGlobalState();
   const [{ loading: loadingArnsState, domains, ants }, dispatchArNSState] =
     useArNSState();
@@ -26,9 +28,9 @@ function Manage() {
   const [search, setSearch] = useState<string>('');
 
   return (
-    <div className="overflow-auto px-[100px] pb-[30px] pt-[10px]">
+    <div className="overflow-auto px-4 md:px-[100px] pb-[30px] pt-[10px]">
       <div className="flex flex-col gap-[10px]">
-        <div className="flex justify-start">
+        <div className="flex justify-between items-center">
           <h1
             className="flex white text-[2rem]"
             style={{
@@ -39,6 +41,12 @@ function Manage() {
           >
             Manage Assets
           </h1>
+          <button
+            onClick={() => navigate('/')}
+            className="whitespace-nowrap rounded-[4px] hidden md:block bg-primary px-4 py-2 text-sm text-black font-medium transition-all hover:bg-primary-dark hover:scale-105"
+          >
+            Register a Name
+          </button>
         </div>
         <div
           id="manage-table-wrapper"
@@ -86,15 +94,17 @@ function Manage() {
                               payload: { showUpgradeAntModal: true },
                             })
                           }
-                          className="h-fit animate-pulse whitespace-nowrap rounded-[4px] bg-primary-thin px-4 py-1 text-sm text-primary transition-all hover:bg-primary hover:text-black"
+                          className="h-fit animate-pulse whitespace-nowrap rounded-[4px] bg-primary-thin px-4 py-1 text-sm text-primary transition-all hover:bg-primary hover:text-black hover:scale-105"
                         >
-                          Upgrade Domains
+                          Upgrade ANTs
                         </button>
                       }
                     />
                   )}
                 <button
-                  className={'button center pointer'}
+                  className={
+                    'button center pointer transition-transform duration-200 ease-in-out hover:scale-105'
+                  }
                   onClick={() =>
                     walletAddress
                       ? dispatchArNSUpdate({